ADD opencl test

+++ base_utils/include/image_utils.h
@@ -128,7 +128,7 @@ void draw_rects_texts(cv::Mat &image, vector<cv::Rect> rects, vector<string> tex
  * @param points
  * @param skeleton 需要连接的ID序号
  */
-void draw_lines(cv::Mat &image, vector<cv::Point> points, const vector<vector<float>> skeleton);
+void draw_lines(cv::Mat &image, vector<cv::Point> points, const vector<vector<int>> skeleton);
 
 /***
  * 绘制带箭头的连接线
@@ -136,6 +136,6 @@ void draw_lines(cv::Mat &image, vector<cv::Point> points, const vector<vector<fl
  * @param points
  * @param skeleton 需要连接的ID序号
  */
-void draw_arrowed_lines(cv::Mat &image, vector<cv::Point> points, const vector<vector<float>> skeleton);
+void draw_arrowed_lines(cv::Mat &image, vector<cv::Point> points, const vector<vector<int>> skeleton);
 
 #endif //DETECTOR_IMAGE_UTILS_H
